Creating separate layout component
