+++ shared/meta/src/index.ts
@@ -1,4 +1,4 @@
-export * from './lib/layout/layout';
-export * from './lib/async-wrapper/async-wrapper';
-export { default as AsyncWrapper } from './lib/async-wrapper/async-wrapper';
-export { default as Sitenav } from './lib/sitenav/sitenav';
+export * from "./lib/async-wrapper/async-wrapper";
+export { default as Layout } from "./lib/layout/layout";
+export { default as AsyncWrapper } from "./lib/async-wrapper/async-wrapper";
+export { default as Sitenav } from "./lib/sitenav/sitenav";
