+++ shared/meta/src/lib/layout/layout.styled.tsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const StyledLayout = styled.div`
+  text-align: center;
+  height: 100vh;
+`;
