--- shared/ui-library/src/lib/loader/loader.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import Loader from './loader';
-
-describe('Loader', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<Loader />);
-    expect(baseElement).toBeTruthy();
-  });
-});
